backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
httpx-oauth==0.13.0
resend==0.6.0
itsdangerous==2.1.2
python-multipart==0.0.6
python-dotenv==1.0.0
pydantic-settings==2.1.0
email-validator==2.2.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.24.1
pytest-cov==4.1.0
